style: tweak the them of the Markdown
