+++ assistant/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@petercatai/assistant",
-  "version": "2.0.13-alpha.6",
+  "version": "2.0.14-alpha.0",
   "description": "PeterCat Assistant Application",
   "repository": "https://github.com/petercat-ai/petercat.git",
   "license": "MIT",
@@ -67,6 +67,11 @@
     "dotenv": "^16.4.5",
     "lodash": "^4.17.21",
     "lottie-react": "^2.4.0",
+    "react-markdown": "7.0.0",
+    "react-syntax-highlighter": "15.0.0",
+    "rehype-highlight": "6.0.0",
+    "remark-gfm": "3.0.1",
+    "remark-html": "^16.0.1",
     "swr": "^2.2.5",
     "tailwindcss": "^3.4.1",
     "tailwindcss-scoped-preflight": "^3.2.6"
