+++ assistant/tailwind.css
@@ -202,34 +202,23 @@
     border: 0;
   }
 
-  p {
-    line-height: 32px;
-  }
-
-  p,
-  pre {
-    margin: 0.6em 0;
-  }
-
-  pre > code[class*='language-'] {
-    line-height: 1.8;
-  }
-
   pre[class*='language-'] {
     background-color: rgba(0, 0, 0, 0.03);
   }
+
   code {
     margin: 0 0.2em;
     padding-inline: 0.4em;
     padding-block: 0.2em 0.1em;
     font-family: 'SFMono-Regular', Consolas, 'Liberation Mono', Menlo, Courier, monospace;
-    background: rgba(150, 150, 150, 0.1);
+    color: #d56161;
     border: 1px solid rgba(100, 100, 100, 0.2);
-    border-radius: 3px;
+ 
   }
 
   code[class*='language-'] {
     background: none;
+    border: none;
   }
 
   code[class*='language-text'] {
