+++ client/package.json
@@ -24,7 +24,7 @@
     "@fullpage/react-fullpage": "^0.1.42",
     "@next/bundle-analyzer": "^13.4.19",
     "@nextui-org/react": "^2.2.9",
-    "@petercatai/assistant": "^2.0.11",
+    "@petercatai/assistant": "^2.0.13",
     "@sentry/nextjs": "^8.28.0",
     "@supabase/supabase-js": "^2.32.0",
     "@tanstack/react-query": "^5.17.19",
